refactor(draft-stage): split DraftStage into main + sidebar + orchestrator
Extracts DraftStageMain.svelte (owns the review controller + the
DraftingDesk column + handleCompleteScene) and DraftStageSidebar.svelte
(owns activeTab, metrics derivations, and the five tab panels,
including the handleVerifyIR / handleUpdateIR handlers it delegates
to IRInspector). DraftStage.svelte shrinks from 650 to ~105 lines
and now only composes SceneSequencer + Main + Sidebar + SceneAuthoringModal
and holds the top-level canGenerate / gateMessages $derived blocks.
activeTab is bound between DraftStage and Sidebar so the Main column
can still flip it to "ir" via onOpenIRTab.

Part of #34.
